adds examples to code component
