+++ client/components/Annyang.js
@@ -63,6 +63,7 @@ export class Annyang extends React.Component {
           <h3>Your challenge is {name}</h3>
           <p>As a reminder, the prompt is: </p>
           <p id="prompt">{prompt}</p>
+
           <p>Now say how you'd solve the problem!</p>
           <button onClick={this.annyangStart}>Start Recording</button>
           <button onClick={this.annyangStop}>Stop Recording</button>
