+++ client/components/Examples.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
-import {getAgent} from '../store/chatbot'
 
 class Examples extends React.Component {
   constructor() {
@@ -9,9 +8,6 @@ class Examples extends React.Component {
     this.state = {
       examples: ''
     }
-  }
-  componentDidMount() {
-    this.props.getAgent()
   }
 
   render() {
@@ -44,11 +40,6 @@ class Examples extends React.Component {
             </p>
           </div>
         </div>
-        {this.props.agent.example ? (
-          <p>Your example was: {this.props.agent.example}</p>
-        ) : (
-          <p />
-        )}
 
         <div />
 
@@ -59,10 +50,7 @@ class Examples extends React.Component {
 }
 
 const mapState = state => ({
-  challenge: state.problems.selected,
-  agent: state.agent
+  challenge: state.problems.selected
 })
 
-const mapDispatch = {getAgent}
-
-export default connect(mapState, mapDispatch)(Examples)
+export default connect(mapState)(Examples)
