+++ client/components/codeView.js
@@ -1,38 +1,52 @@
 import React from 'react'
 import {ConnectedSandbox} from './codeSandbox'
-import {PromiseProvider} from 'mongoose'
-import {connect} from 'react-redux'
 
-export const CodeView = props => {
-  return (
-    <div id="code" className="content">
-      <div className="challengeDetails">
-        <div>Your Challenge: {props.problem.name}</div>
-        <br />
-        <div>Prompt: {props.problem.prompt}</div>
-      </div>
-      <div id="sandbox">
-        <ConnectedSandbox />
-      </div>
-      <div className="btnContainer">
-        <div className="nextBtn" onClick={() => props.history.push('/stats')}>
-          Next Step - Check Your Stats!
+import {connect} from 'react-redux'
+import {getAgent} from '../store/chatbot'
+
+class CodeView extends React.Component {
+  componentDidMount() {
+    this.props.getAgent()
+  }
+  render() {
+    return (
+      <div id="code" className="content">
+        <div className="challengeDetails">
+          <div>Your Challenge: {this.props.problem.name}</div>
+          <br />
+          <div>Prompt: {this.props.problem.prompt}</div>
+        </div>
+        {this.props.agent.example ? (
+          <p>Your example was: {this.props.agent.example}</p>
+        ) : (
+          <p />
+        )}
+
+        <div id="sandbox">
+          <ConnectedSandbox />
+        </div>
+        <div className="btnContainer">
+          <div
+            className="nextBtn"
+            onClick={() => this.props.history.push('/stats')}
+          >
+            Next Step - Check Your Stats!
+          </div>
         </div>
       </div>
-    </div>
-  )
+    )
+  }
 }
 
 const mapState = state => {
   return {
-    problem: state.problems.selected
+    problem: state.problems.selected,
+    agent: state.agent
   }
 }
 
-// const mapDispatch = dispatch => {
+const mapDispatch = {getAgent}
 
-// }
-
-export const Code = connect(mapState, null)(CodeView)
+export const Code = connect(mapState, mapDispatch)(CodeView)
 
 export default Code
