+++ server/api/chatbot.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 
-let agent = {}
+let agent = {hello: 'hi there'}
 
 //Create an instance
 
@@ -9,7 +9,8 @@ module.exports = router
 router.post('/', (req, res, next) => {
   try {
     console.log('req.body in chatbot api', req.body)
-    agent.example = [...agent.example, req.body.queryResult.queryText]
+    agent.example = req.body.queryResult.queryText
+    console.log('agent in the chatbot api', agent)
     res.sendStatus(200)
   } catch (err) {
     next(err)
